Obtener valoraciones closes #25

+++ P4/RecommenderSystem/src/fblupi/rs/movielens/Users.java
@@ -57,7 +57,7 @@ public class Users {
         Iterator userEntries = userRatings.entrySet().iterator();
         while (userEntries.hasNext()) {
             Map.Entry entry = (Map.Entry) userEntries.next();
-            userAverage += (double) entry.getValue();
+            userAverage += (int) entry.getValue();
         }
         userAverage /= userRatings.size();
